di 분리
